+++ src/output/mod.rs
@@ -1,4 +1,5 @@
 pub mod paste;
+pub mod clipget;
 
 #[cfg(target_os = "linux")]
 pub mod clipboard;